+++ libplatune/player/src/dto/player_state.rs
@@ -1,6 +1,6 @@
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, uniffi::Record)]
 pub struct PlayerState {
     pub volume: f64,
     pub queue: Vec<String>,
-    pub queue_position: usize,
+    pub queue_position: u32,
 }
